Move ProductEndpoints to net package
