+++ app/src/main/kotlin/br/com/devhernand/starwars/view/main/MainModule.kt
@@ -2,16 +2,12 @@ package br.com.devhernand.starwars.view.main
 
 import android.arch.lifecycle.ViewModelProvider
 import br.com.devhernand.starwars.ViewModelProviderFactory
-import br.com.devhernand.starwars.data.repository.datasource.ProductEndpoints
 import br.com.devhernand.starwars.di.SchedulerProvider
-import br.com.devhernand.starwars.di.app.NetworkModule
 import br.com.devhernand.starwars.domain.api.br.com.devhernand.starwars.domain.usecases.MainUseCase
 import br.com.devhernand.starwars.domain.api.br.com.devhernand.starwars.domain.usecases.MainUseCaseImpl
 import br.com.devhernand.starwars.domain.api.br.com.devhernand.starwars.domain.repository.ProductRepository
 import dagger.Module
 import dagger.Provides
-import retrofit2.Retrofit
-import javax.inject.Singleton
 
 /**
  * Created by Nando on 23/12/2017.
